+++ src/jsx-info.ts
@@ -1,0 +1,13 @@
+#!/usr/bin/env node
+import chalk from "chalk";
+import { main } from "./main";
+import { printError } from "./printer";
+
+main().catch((err) => {
+  if (process.env.DEBUG) {
+    printError(err);
+  } else {
+    printError(chalk.bold.red(err.message));
+  }
+  process.exit(1);
+});